@@ -26,7 +26,7 @@
         }
     },
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "symfony/ux-twig-component": "^1.4"
     },
     "require-dev": {
@@ -40,7 +40,8 @@
         "doctrine/doctrine-bundle": "^2.0",
         "doctrine/orm": "^2.7",
         "zenstruck/foundry": "^1.10",
-        "zenstruck/browser": "^0.5.0"
+        "zenstruck/browser": "^0.5.0",
+        "doctrine/annotations": "^1.0"
     },
     "extra": {
         "branch-alias": {
