@@ -12,21 +12,21 @@
 namespace Symfony\UX\LiveComponent\Tests\Fixture\Component;
 
 use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
 use Symfony\UX\LiveComponent\Attribute\BeforeReRender;
 use Symfony\UX\LiveComponent\Attribute\LiveAction;
 use Symfony\UX\LiveComponent\Attribute\LiveProp;
 use Symfony\UX\LiveComponent\Attribute\PostHydrate;
 use Symfony\UX\LiveComponent\Attribute\PreDehydrate;
-use Symfony\UX\LiveComponent\LiveComponentInterface;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class Component2 implements LiveComponentInterface
+#[AsLiveComponent('component2')]
+final class Component2
 {
-    /**
-     * @LiveProp
-     */
+    #[LiveProp]
     public int $count = 1;
 
     public bool $preDehydrateCalled = false;
@@ -35,46 +35,31 @@ final class Component2 implements LiveComponentInterface
 
     public bool $beforeReRenderCalled = false;
 
-    /**
-     * @LiveAction
-     */
+    #[LiveAction]
     public function increase(): void
     {
         ++$this->count;
     }
 
-    /**
-     * @LiveAction
-     */
-    public function redirect(): RedirectResponse
+    #[LiveAction]
+    public function redirect(UrlGeneratorInterface $urlGenerator): RedirectResponse
     {
-        return new RedirectResponse('/');
+        return new RedirectResponse($urlGenerator->generate('homepage'));
     }
 
-    public static function getComponentName(): string
-    {
-        return 'component2';
-    }
-
-    /**
-     * @PreDehydrate()
-     */
+    #[PreDehydrate]
     public function preDehydrateMethod(): void
     {
         $this->preDehydrateCalled = true;
     }
 
-    /**
-     * @PostHydrate()
-     */
+    #[PostHydrate]
     public function postHydrateMethod(): void
     {
         $this->postHydrateCalled = true;
     }
 
-    /**
-     * @BeforeReRender()
-     */
+    #[BeforeReRender]
     public function beforeReRenderMethod(): void
     {
         $this->beforeReRenderCalled = true;
