@@ -11,28 +11,20 @@
 
 namespace Symfony\UX\LiveComponent\Tests\Fixture\Component;
 
+use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
 use Symfony\UX\LiveComponent\Attribute\LiveProp;
-use Symfony\UX\LiveComponent\LiveComponentInterface;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
  */
-final class Component3 implements LiveComponentInterface
+#[AsLiveComponent('component3')]
+final class Component3
 {
-    /**
-     * @LiveProp(fieldName="myProp1")
-     */
+    #[LiveProp(fieldName: 'myProp1')]
     public $prop1;
 
-    /**
-     * @LiveProp(fieldName="getProp2Name()")
-     */
+    #[LiveProp(fieldName: 'getProp2Name()')]
     public $prop2;
-
-    public static function getComponentName(): string
-    {
-        return 'component_3';
-    }
 
     public function getProp2Name(): string
     {
