@@ -13,7 +13,6 @@ namespace Symfony\UX\LiveComponent\Tests\Unit\Attribute;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\UX\LiveComponent\Attribute\LiveProp;
-use Symfony\UX\LiveComponent\LiveComponentInterface;
 
 /**
  * @author Kevin Bond <kevinbond@gmail.com>
@@ -22,54 +21,39 @@ final class LivePropTest extends TestCase
 {
     public function testHydrateWithMethod(): void
     {
-        $this->assertSame('someMethod', (new LiveProp(['hydrateWith' => 'someMethod']))->hydrateMethod());
-        $this->assertSame('someMethod', (new LiveProp(['hydrateWith' => 'someMethod()']))->hydrateMethod());
+        $this->assertSame('someMethod', (new LiveProp(false, [], 'someMethod'))->hydrateMethod());
+        $this->assertSame('someMethod', (new LiveProp(false, [], 'someMethod()'))->hydrateMethod());
     }
 
     public function testDehydrateWithMethod(): void
     {
-        $this->assertSame('someMethod', (new LiveProp(['dehydrateWith' => 'someMethod']))->dehydrateMethod());
-        $this->assertSame('someMethod', (new LiveProp(['dehydrateWith' => 'someMethod()']))->dehydrateMethod());
+        $this->assertSame('someMethod', (new LiveProp(false, [], null, 'someMethod'))->dehydrateMethod());
+        $this->assertSame('someMethod', (new LiveProp(false, [], null, 'someMethod()'))->dehydrateMethod());
     }
 
     public function testCanCallCalculateFieldNameAsString(): void
     {
-        $component = new class() implements LiveComponentInterface {
-            public static function getComponentName(): string
-            {
-                return 'name';
-            }
-        };
+        $component = new class() {};
 
-        $this->assertSame('field', (new LiveProp(['fieldName' => 'field']))->calculateFieldName($component, 'fallback'));
+        $this->assertSame('field', (new LiveProp(false, [], null, null, 'field'))->calculateFieldName($component, 'fallback'));
     }
 
     public function testCanCallCalculateFieldNameAsMethod(): void
     {
-        $component = new class() implements LiveComponentInterface {
-            public static function getComponentName(): string
-            {
-                return 'name';
-            }
-
+        $component = new class() {
             public function fieldName(): string
             {
                 return 'foo';
             }
         };
 
-        $this->assertSame('foo', (new LiveProp(['fieldName' => 'fieldName()']))->calculateFieldName($component, 'fallback'));
+        $this->assertSame('foo', (new LiveProp(false, [], null, null, 'fieldName()'))->calculateFieldName($component, 'fallback'));
     }
 
     public function testCanCallCalculateFieldNameWhenNotSet(): void
     {
-        $component = new class() implements LiveComponentInterface {
-            public static function getComponentName(): string
-            {
-                return 'name';
-            }
-        };
+        $component = new class() {};
 
-        $this->assertSame('fallback', (new LiveProp([]))->calculateFieldName($component, 'fallback'));
+        $this->assertSame('fallback', (new LiveProp())->calculateFieldName($component, 'fallback'));
     }
 }
